tests/wpt/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssUrlImageValue.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>CSSURLImageValue tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#imagevalue-objects">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/testhelper.js"></script>
<body>
<script>
'use strict';

const gTestUrl = '../resources/1x1-green.png';
const gBase64TestUrl = 'data:image/gif;base64,R0lGODlhAQABAAAAACwAAAAAAQABAAA=';
const gBadTestUrl = document.location.href;

test(() => {
  assert_throws(new TypeError(), () => new CSSURLImageValue("file://:This is invalid url"));
}, 'Constructing a CSSURLImageValue with an invalid URL throws a TypeError');

test(() => {
  const result = new CSSURLImageValue(gTestUrl);
  assert_equals(result.url, gTestUrl);
  assert_equals(result.intrinsicWidth, null);
  assert_equals(result.intrinsicHeight, null);
  assert_equals(result.intrinsicRatio, null);
  assert_equals(result.state, 'unloaded');
}, 'Constructing a CSSURLImageValue with a valid URL puts it in an unloaded state');

test(() => {
  let result = new CSSURLImageValue(gTestUrl);
  assert_throws(new TypeError(), () => result.url = gBase64TestUrl);
}, 'CSSURLImageValue.url is readonly');

async_test(t => {
  const result = new CSSURLImageValue(gTestUrl);
  let image = loadImageResource(t, result);

  image.addEventListener('load', t.step_func_done(() => {
    assert_equals(result.url, gTestUrl);
    assert_equals(result.state, 'loaded');
    assert_equals(result.intrinsicWidth, 1);
    assert_equals(result.intrinsicHeight, 1);
    assert_equals(result.intrinsicRatio, 1);
  }));
}, 'Loading a CSSURLImageValue from a URL sets its state to loaded');

async_test(t => {
  const result = new CSSURLImageValue(gBase64TestUrl);
  let image = loadImageResource(t, result);

  image.addEventListener('load', t.step_func_done(() => {
    assert_equals(result.url, gBase64TestUrl);
    assert_equals(result.state, 'loaded');
    assert_equals(result.intrinsicWidth, 1);
    assert_equals(result.intrinsicHeight, 1);
    assert_equals(result.intrinsicRatio, 1);
  }));
}, 'Loading a CSSURLImageValue from a base64 URL sets its state to loaded');

async_test(t => {
  const result = new CSSURLImageValue(gBadTestUrl);
  let image = loadImageResource(t, result);

  image.addEventListener('error', t.step_func_done(() => {
      assert_equals(result.url, gBadTestUrl);
      assert_equals(result.state, 'error');
      assert_equals(result.intrinsicWidth, null);
      assert_equals(result.intrinsicHeight, null);
      assert_equals(result.intrinsicRatio, null);
  }));
}, 'Loading a CSSURLImageValue from a URL to an invalid image sets its state to error');

</script>
